script.js: Drop window.onresize
window.onresize should be avoided if possible, especially when it's only
used for stylesheet manipulation that could be done with CSS.

In addition, setTimeout should be avoided especially when it's only used
for chaining transitions. Use transition delays instead.

Now, the collapsed menu is hidden by default instead of relying on
JavaScript to "hide" the menu by default.